feat: Add methods to add single or multiple layers to DynamicLogger
This commit adds two methods to add a single layer and multiple layers to `DynamicLogger`:
- `add_layer` to add a single layer to the logger
- `add_layers` to add multiple layers to the logger

These methods take in layer(s) implementing the `Layer` trait for
`Registry` and push them into the existing list of layers in
`DynamicLogger`.

This feature will make it easier to add layers dynamically during
run-time.