feat(error): add custom error types and improve error handling
- Added a new module 'error' to define custom error types.
- Added a new enum 'DynLogAPIErr' to define different error variants.
- Improved error handling by adding more descriptive error messages and by returning custom error types.
- Renamed 'initialize_logger' to 'init'.
- Refactored register_filelogger_target to return 'DynLogAPIErr' instead of 'Result<()>'

Note: This commit also adds a new dependency 'thiserror' and removes 'anyhow' in Cargo.toml file.

The new error module defines a custom 'DynLogAPIErr' enum which can help
us to better understand the source of errors. Additionally, this commit
refactors the init function to improve error handling.

Previously, if there was any error while reading the log file, it would
result in a panic. Now, we wrap the errors in a custom `DynLogAPIErr`
variant, which makes it more clear what kind of error we encountered.